Második verzió, szétszedve
A meglévő kódot szétszedtem egy új Main.py-ra és egy Models.py-ra
